Add character.js and cleared (seperated) the code
