+++ index.html
@@ -34,7 +34,9 @@
       </button>
     </header>
     <main class="grid-main"></main>
-    <footer class="grid-footer"></footer>
+    <footer class="grid-footer">
+      <p>with Love from Alex</p>
+    </footer>
     <script type="module" src="/src/main.js"></script>
   </body>
 </html>
