+++ src/lib/API.js
@@ -1,21 +1,9 @@
+import { createSection } from "./character";
+
 export function getCharacter() {
-  let url;
-  const filter = document.querySelector(".header__filter");
-  const status = filter.value;
-  const inputName = document.querySelector(".header__search-form");
-  const searchName = inputName.value;
-
-  if (status == "Alive") {
-    url = `https://rickandmortyapi.com/api/character?name=${searchName}&status=alive`;
-  } else if (status == "Dead") {
-    url = `https://rickandmortyapi.com/api/character?name=${searchName}&status=dead`;
-  } else if (status == "Unknown") {
-    url = `https://rickandmortyapi.com/api/character?name=${searchName}&status=unknown`;
-  } else if (status == "All") {
-    url = `https://rickandmortyapi.com/api/character?name=${searchName}`;
-  }
-
-  console.log(url);
+  const filterDrop = document.querySelector(".header__filter");
+  const status = filterDrop.value;
+  let url = createUrlFrom(status);
 
   fetch(url)
     .then((response) => {
@@ -27,30 +15,26 @@ export function getCharacter() {
     })
     .then((dataObject) =>
       dataObject.results.forEach((character) => {
-        const main = document.querySelector("main");
-        const section = document.createElement("section");
-        section.classList.add("character");
-        const fullName = document.createElement("h2");
-        fullName.classList.add("fullName");
-        fullName.textContent = `${character.name}`;
-        const img = document.createElement("img");
-        img.classList.add("img");
-        img.src = character.image;
-        img.alt = `$(character.name)`;
-
-        if (character.status === "Alive") {
-          section.style.backgroundColor = "var(--section-alive-color)";
-        } else if (character.status === "Dead") {
-          section.style.backgroundColor = "var(--section-dead-color)";
-        } else if (character.status === "unknown") {
-          section.style.backgroundColor = "var(--section-unknown-color)";
-        }
-
-        section.append(fullName);
-        section.append(img);
-        main.append(section);
+        createSection(character);
       })
     );
+}
+
+function createUrlFrom(status) {
+  let url;
+  const baseUrl = "https://rickandmortyapi.com/api/character/";
+  const inputName = document.querySelector(".header__search-form");
+  const searchName = inputName.value;
+  if (status == "Alive") {
+    url = `${baseUrl}?name=${searchName}&status=alive`;
+  } else if (status == "Dead") {
+    url = `${baseUrl}?name=${searchName}&status=dead`;
+  } else if (status == "Unknown") {
+    url = `${baseUrl}?name=${searchName}&status=unknown`;
+  } else if (status == "All") {
+    url = `${baseUrl}?name=${searchName}`;
+  }
+  return url;
 }
 
 export function clearcharacters() {
