+++ src/main.js
@@ -8,3 +8,11 @@ button.addEventListener("click", () => {
   clearcharacters();
   getCharacter();
 });
+
+const form = document.querySelector("form");
+
+form.addEventListener("submit", (event) => {
+  event.preventDefault();
+  clearCharacters();
+  createCharacter();
+});
